Element: Reorganize code to make it fit better React Native needs
