+++ packages/element/src/index.js
@@ -1,197 +1,7 @@
-/**
- * External dependencies
- */
-import {
-	createElement,
-	createContext,
-	createRef,
-	forwardRef,
-	Component,
-	cloneElement,
-	Children,
-	Fragment,
-	isValidElement,
-	StrictMode,
-} from 'react';
-import { render, findDOMNode, createPortal, unmountComponentAtNode } from 'react-dom';
-import { isString } from 'lodash';
+export * from './react';
+export * from './react-platform';
+export { default as renderToString } from './serialize';
+export { default as RawHTML } from './raw-html';
 
-/**
- * Internal dependencies
- */
-import serialize from './serialize';
-import { RawHTML } from './index-common';
+// deprecated
 export * from './deprecated';
-
-/**
- * Returns a new element of given type. Type can be either a string tag name or
- * another function which itself returns an element.
- *
- * @param {?(string|Function)} type     Tag name or element creator
- * @param {Object}             props    Element properties, either attribute
- *                                       set to apply to DOM node or values to
- *                                       pass through to element creator
- * @param {...WPElement}       children Descendant elements
- *
- * @return {WPElement} Element.
- */
-export { createElement };
-
-/**
- * Returns an object tracking a reference to a rendered element via its
- * `current` property as either a DOMElement or Element, dependent upon the
- * type of element rendered with the ref attribute.
- *
- * @return {Object} Ref object.
- */
-export { createRef };
-
-/**
- * Component enhancer used to enable passing a ref to its wrapped component.
- * Pass a function argument which receives `props` and `ref` as its arguments,
- * returning an element using the forwarded ref. The return value is a new
- * component which forwards its ref.
- *
- * @param {Function} forwarder Function passed `props` and `ref`, expected to
- *                             return an element.
- *
- * @return {WPComponent} Enhanced component.
- */
-export { forwardRef };
-
-/**
- * Renders a given element into the target DOM node.
- *
- * @param {WPElement} element Element to render
- * @param {Element}   target  DOM node into which element should be rendered
- */
-export { render };
-
-/**
- * Removes any mounted element from the target DOM node.
- *
- * @param {Element} target DOM node in which element is to be removed
- */
-export { unmountComponentAtNode };
-
-/**
- * A base class to create WordPress Components (Refs, state and lifecycle hooks)
- */
-export { Component };
-
-/**
- * Creates a copy of an element with extended props.
- *
- * @param {WPElement} element Element
- * @param {?Object}   props   Props to apply to cloned element
- *
- * @return {WPElement} Cloned element.
- */
-export { cloneElement };
-
-/**
- * Finds the dom node of a React component
- *
- * @param {Component} component component's instance
- * @param {Element}   target    DOM node into which element should be rendered
- */
-export { findDOMNode };
-
-export { Children };
-
-export { StrictMode };
-
-/**
- * A component which renders its children without any wrapping element.
- */
-export { Fragment };
-
-/**
- * Creates a context object containing two components: a provider and consumer.
- *
- * @param {Object} defaultValue A default data stored in the context.
- *
- * @return {Object} Context object.
- */
-export { createContext };
-
-/**
- * Checks if an object is a valid WPElement
- *
- * @param {Object} objectToCheck The object to be checked.
- *
- * @return {boolean} true if objectToTest is a valid WPElement and false otherwise.
- */
-export { isValidElement };
-
-/**
- * Creates a portal into which a component can be rendered.
- *
- * @see https://github.com/facebook/react/issues/10309#issuecomment-318433235
- *
- * @param {Component} component Component
- * @param {Element}   target    DOM node into which element should be rendered
- */
-export { createPortal };
-
-/**
- * Renders a given element into a string.
- *
- * @param {WPElement} element Element to render
- *
- * @return {string} HTML.
- */
-export { serialize as renderToString };
-
-/**
- * Concatenate two or more React children objects.
- *
- * @param {...?Object} childrenArguments Array of children arguments (array of arrays/strings/objects) to concatenate.
- *
- * @return {Array} The concatenated value.
- */
-export function concatChildren( ...childrenArguments ) {
-	return childrenArguments.reduce( ( memo, children, i ) => {
-		Children.forEach( children, ( child, j ) => {
-			if ( child && 'string' !== typeof child ) {
-				child = cloneElement( child, {
-					key: [ i, j ].join(),
-				} );
-			}
-
-			memo.push( child );
-		} );
-
-		return memo;
-	}, [] );
-}
-
-/**
- * Switches the nodeName of all the elements in the children object.
- *
- * @param {?Object} children Children object.
- * @param {string}  nodeName Node name.
- *
- * @return {?Object} The updated children object.
- */
-export function switchChildrenNodeName( children, nodeName ) {
-	return children && Children.map( children, ( elt, index ) => {
-		if ( isString( elt ) ) {
-			return createElement( nodeName, { key: index }, elt );
-		}
-		const { children: childrenProp, ...props } = elt.props;
-		return createElement( nodeName, { key: index, ...props }, childrenProp );
-	} );
-}
-
-/**
- * Component used as equivalent of Fragment with unescaped HTML, in cases where
- * it is desirable to render dangerous HTML without needing a wrapper element.
- * To preserve additional props, a `div` wrapper _will_ be created if any props
- * aside from `children` are passed.
- *
- * @param {string} props.children HTML to render.
- *
- * @return {WPElement} Dangerously-rendering element.
- */
-export { RawHTML };
