--- packages/element/src/index.native.js
@@ -1,53 +1,0 @@
-/**
- * External dependencies
- */
-import {
-	createElement,
-	Component,
-} from 'react';
-
-/**
- * Internal dependencies
- */
-import serialize from './serialize';
-import { RawHTML } from './index-common';
-
-/**
- * Returns a new element of given type. Type can be either a string tag name or
- * another function which itself returns an element.
- *
- * @param {?(string|Function)} type     Tag name or element creator
- * @param {Object}             props    Element properties, either attribute
- *                                       set to apply to DOM node or values to
- *                                       pass through to element creator
- * @param {...WPElement}       children Descendant elements
- *
- * @return {WPElement} Element.
- */
-export { createElement };
-
-/**
- * A base class to create WordPress Components (Refs, state and lifecycle hooks)
- */
-export { Component };
-
-/**
- * Renders a given element into a string.
- *
- * @param {WPElement} element Element to render
- *
- * @return {string} HTML.
- */
-export { serialize as renderToString };
-
-/**
- * Component used as equivalent of Fragment with unescaped HTML, in cases where
- * it is desirable to render dangerous HTML without needing a wrapper element.
- * To preserve additional props, a `div` wrapper _will_ be created if any props
- * aside from `children` are passed.
- *
- * @param {string} props.children HTML to render.
- *
- * @return {WPElement} Dangerously-rendering element.
- */
-export { RawHTML };
