+++ packages/element/src/react-platform.js
@@ -1,0 +1,42 @@
+/**
+ * External dependencies
+ */
+import {
+	createPortal,
+	findDOMNode,
+	render,
+	unmountComponentAtNode,
+} from 'react-dom';
+
+/**
+ * Creates a portal into which a component can be rendered.
+ *
+ * @see https://github.com/facebook/react/issues/10309#issuecomment-318433235
+ *
+ * @param {Component} component Component
+ * @param {Element}   target    DOM node into which element should be rendered
+ */
+export { createPortal };
+
+/**
+ * Finds the dom node of a React component
+ *
+ * @param {Component} component component's instance
+ * @param {Element}   target    DOM node into which element should be rendered
+ */
+export { findDOMNode };
+
+/**
+ * Renders a given element into the target DOM node.
+ *
+ * @param {WPElement} element Element to render
+ * @param {Element}   target  DOM node into which element should be rendered
+ */
+export { render };
+
+/**
+ * Removes any mounted element from the target DOM node.
+ *
+ * @param {Element} target DOM node in which element is to be removed
+ */
+export { unmountComponentAtNode };
