+++ packages/element/src/test/index.js
@@ -1,17 +1,12 @@
 /**
- * External dependencies
- */
-import { shallow } from 'enzyme';
-
-/**
  * Internal dependencies
  */
 import {
-	createElement,
 	concatChildren,
+	createElement,
+	RawHTML,
 	renderToString,
 	switchChildrenNodeName,
-	RawHTML,
 } from '../';
 
 describe( 'element', () => {
@@ -118,35 +113,6 @@ describe( 'element', () => {
 			expect( children[ 0 ].props.align ).toBe( 'left' );
 			expect( children[ 1 ].type ).toBe( 'em' );
 			expect( children[ 1 ].props.children ).toBe( 'Concombre' );
-		} );
-	} );
-
-	describe( 'RawHTML', () => {
-		it( 'is dangerous', () => {
-			const html = '<p>So scary!</p>';
-			const element = shallow(
-				<RawHTML>
-					{ html }
-				</RawHTML>
-			);
-
-			expect( element.type() ).toBe( 'div' );
-			expect( element.prop( 'dangerouslySetInnerHTML' ).__html ).toBe( html );
-			expect( element.prop( 'children' ) ).toBe( undefined );
-		} );
-
-		it( 'creates wrapper if assigned other props', () => {
-			const html = '<p>So scary!</p>';
-			const element = shallow(
-				<RawHTML className="foo">
-					{ html }
-				</RawHTML>
-			);
-
-			expect( element.type() ).toBe( 'div' );
-			expect( element.prop( 'className' ) ).toBe( 'foo' );
-			expect( element.prop( 'dangerouslySetInnerHTML' ).__html ).toBe( html );
-			expect( element.prop( 'children' ) ).toBe( undefined );
 		} );
 	} );
 } );
